backend/requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
django-redis
redis
pillow
gunicorn
boto3
django-storages
google-generativeai
pymongo[srv]==3.12